+++ migrations/1748191419877-planner-category-notes.ts
@@ -1,0 +1,23 @@
+import { getDb } from '../migrations-utils/db';
+
+export const up = async () => {
+  const db = await getDb();
+  db.collection("categories").updateMany({}, [
+    {
+      $set: {
+        notes: []
+      }
+    }
+   ])
+};
+
+export const down = async () => {
+  const db = await getDb();
+  db.collection("categories").updateMany({}, [
+      {
+         $unset: {
+            notes: ""
+         }
+      }
+   ])
+};